+++ hardware/YouLess.cpp
@@ -202,7 +202,7 @@ void CYouLess::GetMeterDetails()
 		Log(LOG_ERROR,"Error connecting to: %s", m_szIPAddress.c_str());
 		return;
 	}
-	int fpos;
+	size_t fpos;
 	std::string pusage=stdstring_trim(results[0]);
 	fpos = pusage.find_first_of(' ');
 	if (fpos!=std::string::npos)
